#7 Cleaning up tests and adding a common project for things shared between tests.
