+++ SpotWelder.IntegrationTests/PostgresTests.cs
@@ -1,0 +1,7 @@
+﻿namespace SpotWelder.IntegrationTests
+{
+  public class PostgresTests 
+    : SpotWelderTestBase
+  {
+  }
+}
