+++ SpotWelder.IntegrationTests/SpotWelder.IntegrationTests.csproj
@@ -11,6 +11,7 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\SpotWelder.Lib\SpotWelder.Lib.csproj" />
+    <ProjectReference Include="..\SpotWelder.Tests.Common\SpotWelder.Tests.Common.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Castle.Core" Version="4.4.0" />
