+++ SpotWelder.IntegrationTests/SpotWelderTestBase.cs
@@ -1,0 +1,10 @@
+﻿using SpotWelder.Tests.Common;
+
+namespace SpotWelder.IntegrationTests
+{
+  public abstract class SpotWelderTestBase
+    : TestBase
+  {
+
+  }
+}
